+++ src/python_ms_core/core/logger/local_logger.py
@@ -1,15 +1,13 @@
-from .providers.azure_logger_config import AzureLoggerConfig
-from ..queue.models.queue_message import QueueMessage
-from ..queue.queue import Queue
 from .abstracts.logger_abstract import LoggerAbstract
+from ..queue.models.queue_message import QueueMessage
+from ..queue.local_queue import LocalQueue
 
 
 class LocalLogger(LoggerAbstract):
 
-    def __init__(self, provider_config=None):
-        super().__init__(provider_config=provider_config)
-        self.config = AzureLoggerConfig(provider_config=provider_config)
-        self.queue_client = Queue(self.config)
+    def __init__(self, config=None):
+        super().__init__(config=config)
+        self.queue_client = LocalQueue(config)
 
     def add_request(self, request_data):
         message = QueueMessage.data_from({
@@ -17,35 +15,35 @@ class LocalLogger(LoggerAbstract):
             'messageType': 'addRequest',
             'data': request_data
         })
-        self.queue_client.send_local(message)
+        self.queue_client.send(message)
 
     def info(self, message: str):
         msg = QueueMessage.data_from({
             'message': message,
             'messageType': 'info',
         })
-        self.queue_client.send_local(msg)
+        self.queue_client.send(msg)
 
     def debug(self, message: str):
         msg = QueueMessage.data_from({
             'message': message,
             'messageType': 'debug',
         })
-        self.queue_client.send_local(msg)
+        self.queue_client.send(msg)
 
     def warn(self, message: str):
         msg = QueueMessage.data_from({
             'message': message,
             'messageType': 'warn',
         })
-        self.queue_client.send_local(msg)
+        self.queue_client.send(msg)
 
     def error(self, message: str):
         msg = QueueMessage.data_from({
             'message': message,
             'messageType': 'error',
         })
-        self.queue_client.send_local(msg)
+        self.queue_client.send(msg)
 
     def record_metric(self, name: str, value: str):
         message = QueueMessage.data_from({
@@ -56,4 +54,4 @@ class LocalLogger(LoggerAbstract):
                 'value': value
             }
         })
-        self.queue_client.send_local(message)
+        self.queue_client.send(message)