+++ CHANGELOG.md
@@ -1,5 +1,30 @@
 # Change log
 
+### 0.0.7
+- Fixed configuration issues
+- Fixed all the review comments
+- Refactor code
+- Removed unused code
+- Added local dev env using TDEI-local-server Package
+- Added local env support to -
+  - Logger
+  - Queue
+  - Topic
+  - Storage
+
+### 0.0.6
+- Added get_remote_url function in FileEntity class which will return the uploaded file url
+
+### 0.0.5
+- Fixed Storage Blob stream
+
+### 0.0.4
+- Fixed threading
+
+### 0.0.3
+- Removed blocking call from topic subscription
+- Added TODOs and FIXMEs
+
 ### 0.0.2
 - Added filtered exception handling
 - Removed unused packages